pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spoofformer"
version = "1.0.0"
description = "Vision Transformer for Face Anti-Spoofing"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "pillow>=9.0.0",
    "scikit-learn>=1.0.0",
    "tqdm>=4.60.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "wandb>=0.15.0",
    "onnx>=1.14.0",
    "onnxruntime>=1.15.0",
]

[tool.setuptools.packages.find]
where = ["src"]
